arreglos

+++ src/components/gestioEmpresas.vue
@@ -4,21 +4,27 @@
     <v-row>
       <!-- Gráfico vacío a la izquierda -->
       <v-col cols="12" md="6">
-        <!-- Aquí puedes agregar tu gráfico o cualquier otro contenido -->
-        <div>
-          gráfico
-          <!-- Contenido del gráfico o cualquier otro elemento -->
-        </div>
+        
+        <v-card>
+          <v-card-title>Gráfico de Barras</v-card-title>
+          <v-card-text>
+            <v-bar-chart
+              :data="getChartData()"
+              :options="getChartOptions()"
+            ></v-bar-chart>
+          </v-card-text>
+        </v-card>
+        
       </v-col>
 
       <!-- Lista de clientes a la derecha -->
       <v-col cols="12" md="6">
-        <v-card height="auto" width="600" style="
+        <v-card max-width="1000" height="auto" width="auto" style="
             position: absolute;
-            margin-left: 300px;
             margin-top: 25px;
             background-color: rgb(189, 236, 255);
           ">
+
           <!-- Contenedor para el botón y el campo de búsqueda -->
           <div style="
               display: flex;
@@ -26,6 +32,7 @@
               align-items: center;
               margin-left: 15px;
             ">
+
             <!-- Botón para crear empresa -->
             <v-btn color="primary" @click="abrirModalCrearEmpresa">Crear Empresa</v-btn>
 
@@ -40,11 +47,12 @@
             style="margin: 15px; width: auto; height: auto">
             <v-row align="center">
               <v-col>
-                <p style="margin: 20px">
+                <p style="margin: 20px; margin-right: 200px;">
                   {{ empresa.nom_empresa }}
                 </p>
               </v-col>
             </v-row>
+            <!-- BTN para eliminar clientes -->
             <v-btn icon @click="eliminarCliente(empresa)" style="position: absolute; right: 10px; bottom: 10px">
               <v-icon>mdi-delete</v-icon>
             </v-btn>
@@ -79,21 +87,36 @@
                     <div style="width: 48%">
                       <label>Teléfono: </label>
 
-                      <!-- Campo para seleccionar el prefijo -->
-                      <select v-model="prefijoTelefonoEditar" class="select-prefijo">
-                        <!-- Opciones de prefijo telefónico -->
-                      </select>
-
                       <!-- Campo para el número de teléfono -->
                       <input v-model="empresaEditada.telefono" required class="bordered-input" type="tel"
                         @input="filtrarNumeros" maxlength="15" />
                     </div>
                   </div>
 
-                  <!-- DIRECCION -->
+                  <!-- PROVINCIA + CIUDAD -->
+                  <div style="
+                      display: flex;
+                      flex-direction: row;
+                      justify-content: space-between;
+                      margin-top: 15px;
+                    ">
+                    <!-- Provincia y CIUDAD -->
+                    <div style="width: 48%;">
+                      <label>Provincia: </label>
+                      <input v-model="empresaEditada.provincia" required class="bordered-input"
+                        style="height: auto; width: 290px" maxlength="45" />
+                    </div>
+                    <div style="width: 48%">
+                      <label>Ciudad: </label>
+                      <input v-model="empresaEditada.ciudad" required class="bordered-input"
+                        style="height: auto; width: 270px" maxlength="45" />
+                    </div>
+                  </div>
+
+                  <!-- Direccion -->
                   <div style="margin-top: 15px">
                     <label>Dirección: </label>
-                    <v-textarea v-model="empresaEditada.direccion" required class="bordered-input"
+                    <v-textarea v-model="empresaEditada.nombre_calle" required class="bordered-input"
                       style="height: auto; width: 653px" maxlength="245"></v-textarea>
                   </div>
 
@@ -140,11 +163,13 @@
               <v-card-title>Crear Empresa</v-card-title>
               <v-card-text>
                 <form>
+                  <!-- NOMBRE + TEL -->
                   <div style="
                       display: flex;
                       flex-direction: row;
                       justify-content: space-between;
                     ">
+                    <!-- NOPMBRE y TELEFONO -->
                     <div style="width: 48%">
                       <label>Nombre: </label>
                       <input v-model="nuevaEmpresa.nombre" required class="bordered-input"
@@ -157,10 +182,30 @@
                     </div>
                   </div>
 
+                  <!-- PROVINCIA + CIUDAD -->
+                  <div style="
+                      display: flex;
+                      flex-direction: row;
+                      justify-content: space-between;
+                      margin-top: 15px;
+                    ">
+                    <!-- Provincia y CIUDAD -->
+                    <div style="width: 48%;">
+                      <label>Provincia: </label>
+                      <input v-model="nuevaEmpresa.provincia" required class="bordered-input"
+                        style="height: auto; width: 290px" maxlength="45" />
+                    </div>
+                    <div style="width: 48%">
+                      <label>Ciudad: </label>
+                      <input v-model="nuevaEmpresa.ciudad" required class="bordered-input"
+                        style="height: auto; width: 270px" maxlength="45" />
+                    </div>
+                  </div>
+
                   <!-- Dirección -->
                   <div style="margin-top: 15px">
                     <label>Dirección: </label>
-                    <v-textarea v-model="nuevaEmpresa.direccion" required class="bordered-input"
+                    <v-textarea v-model="nuevaEmpresa.nombre_calle" required class="bordered-input"
                       style="height: auto; width: 653px" maxlength="245"></v-textarea>
                   </div>
 
@@ -211,23 +256,29 @@
                   <v-list-item v-if="empresaSeleccionada">
                     <v-list-item>
                       <v-list-item-title>{{
-                        empresaSeleccionada.nom_empresa
-                      }}</v-list-item-title>
-                      <v-list-item-subtitle>{{
-                        empresaSeleccionada.direccion
-                      }}</v-list-item-subtitle>
-                      <v-list-item-subtitle>{{
-                        empresaSeleccionada.numero_teléfono
-                      }}</v-list-item-subtitle>
-                      <v-list-item-subtitle>{{
-                        empresaSeleccionada.correo
-                      }}</v-list-item-subtitle>
-                      <v-list-item-subtitle>{{
-                        empresaSeleccionada.sitio_web
-                      }}</v-list-item-subtitle>
-                      <v-list-item-subtitle>{{
+              empresaSeleccionada.nom_empresa
+            }}</v-list-item-title>
+                      <v-list-item-subtitle><b>Provincia: </b> {{
+                empresaSeleccionada.provincia
+              }}</v-list-item-subtitle>
+                      <v-list-item-subtitle><b>Ciudad: </b> {{
+                empresaSeleccionada.ciudad
+              }}</v-list-item-subtitle>
+                      <v-list-item-subtitle><b>Calle: </b> {{
+                empresaSeleccionada.nombre_calle
+              }}</v-list-item-subtitle>
+                      <v-list-item-subtitle><b>TEL: </b>{{
+                empresaSeleccionada.numero_teléfono
+              }}</v-list-item-subtitle>
+                      <v-list-item-subtitle><b>MAIL: </b>{{
+                empresaSeleccionada.correo
+              }}</v-list-item-subtitle>
+                      <v-list-item-subtitle><b>Web: </b>{{
+                empresaSeleccionada.sitio_web
+              }}</v-list-item-subtitle>
+                      <v-list-item-subtitle><b>Suscripción: </b>{{
                         empresaSeleccionada.plan
-                      }}</v-list-item-subtitle>
+                        }}</v-list-item-subtitle>
                     </v-list-item>
                   </v-list-item>
                 </v-list>
@@ -245,7 +296,7 @@
 
 <script>
 import { insertEmpresa, getEmpresa, updateEmpresa, deleteEmpresa } from "@/services/connectionManager";
-import { socket, state } from "../services/socket";
+//import { BarChart } from 'vuetify/lib';
 
 export default {
   data() {
@@ -253,24 +304,27 @@ export default {
       clienteSeleccionado: null,
       modoEdicion: false,
       filtroEmpresa: "",
-
       mostrarModalEditarEmpresa: false,
       empresaEditada: {
         nombre: "",
-        direccion: "",
+        nombre_calle: "",
         telefono: "",
         correo: "",
         plan: "Standard",
         sitioWeb: "",
+        provincia: "",
+        ciudad: "",
         id: ""
       },
       nuevaEmpresa: {
         nombre: "",
-        direccion: "",
+        nombre_calle: "",
         telefono: "",
         correo: "",
         plan: "Standard",
         sitioWeb: "",
+        provincia: "",
+        ciudad: ""
       },
       prefijoTelefonoEditar: "+1",
       clienteEditado: null,
@@ -279,8 +333,9 @@ export default {
       empresaSeleccionada: null,
       empresas: [],
       empresasEnviar: [],
-      empresaEditadaEnviar:[],
+      empresaEditadaEnviar: [],
       mostrarModalCrear: false,
+      
     };
   },
   computed: {
@@ -307,38 +362,51 @@ export default {
       this.nuevaEmpresa.telefono = filteredInput;
     },
 
+    // EDITAR EMPRESAS
     editarEmpresa(empresa) {
       this.empresaSeleccionada = empresa;
 
-       this.empresaEditada.nombre = empresa.nom_empresa;
-       this.empresaEditada.direccion = empresa.direccion;
-       this.empresaEditada.telefono = empresa.numero_teléfono;
-       this.empresaEditada.correo = empresa.correo;
-       this.empresaEditada.sitioWeb = empresa.sitio_web;
-       this.empresaEditada.plan = empresa.plan;
-       this.empresaEditada.id = empresa.id_empresa;
+      this.empresaEditada.nombre = empresa.nom_empresa;
+      this.empresaEditada.nombre_calle = empresa.nombre_calle;
+      this.empresaEditada.telefono = empresa.numero_teléfono;
+      this.empresaEditada.correo = empresa.correo;
+      this.empresaEditada.sitioWeb = empresa.sitio_web;
+      this.empresaEditada.plan = empresa.plan;
+      this.empresaEditada.provincia = empresa.provincia;
+      this.empresaEditada.ciudad = empresa.ciudad;
+      this.empresaEditada.id = empresa.id_empresa;
 
       // Abre el modal de edición de empresa
       this.mostrarModalEditarEmpresa = true;
     },
 
-    async eliminarCliente(empresa){
-      const id_empresa = empresa.id_empresa;
-      await deleteEmpresa(id_empresa);
-      this.getEmpresas();
+    // ELIMINAR EMPRESA
+    async eliminarCliente(empresa) {
+      const confirmacion = confirm("¿Estás seguro que quieres eliminar la empresa " + empresa.nom_empresa + "?");
+      if (confirmacion) {
+        const id_empresa = empresa.id_empresa;
+        await deleteEmpresa(id_empresa);
+        this.getEmpresas();
+      } else {
+        // Si el usuario cancela, no se realiza ninguna acción
+        return;
+      }
     },
 
+    // GUARDAR EDICION EMPRESA
     async guardarEdicionEmpresa() {
       this.empresaEditadaEnviar.push(
         this.empresaEditada.nombre,
-        this.empresaEditada.direccion,
+        this.empresaEditada.nombre_calle,
         this.empresaEditada.telefono,
         this.empresaEditada.correo,
         this.empresaEditada.sitioWeb,
         this.empresaEditada.plan,
+        this.empresaEditada.provincia,
+        this.empresaEditada.ciudad,
         this.empresaEditada.id
       );
-      console.log(this.empresaEditadaEnviar);
+      // console.log(this.empresaEditadaEnviar);
       await updateEmpresa(this.empresaEditadaEnviar)
       this.getEmpresas();
       this.mostrarModalEditarEmpresa = false;
@@ -348,30 +416,34 @@ export default {
     cancelarEdicionEmpresa() {
       // Restablece los valores originales antes de la edición
       this.empresaEditada.nombre = "";
-      this.empresaEditada.direccion = "";
+      this.empresaEditada.nombre_calle = "";
       this.empresaEditada.telefono = "";
       this.empresaEditada.correo = "";
       this.empresaEditada.plan = "Standard";
       this.empresaEditada.sitioWeb = "";
+      this.empresaEditada.provincia = "";
+      this.empresaEditada.ciudad = "";
 
       // Cierra el modal de edición
       this.mostrarModalEditarEmpresa = false;
     },
 
     mostrarFormularioEmpresa() {
-      // Mostrar el modal o sección para crear empresa
+      // Mostrar el modal editar empresa
       this.mostrarModalEditarEmpresa = true;
     },
 
-    // CREAR Y GUARDAAR EMPRESA
+    // CREAR Y GUARDAR EMPRESA
     async crearYGuardarEmpresa() {
       // Validar que todos los campos estén rellenados
       if (
         !this.nuevaEmpresa.nombre ||
-        !this.nuevaEmpresa.direccion ||
+        !this.nuevaEmpresa.nombre_calle ||
         !this.nuevaEmpresa.telefono ||
         !this.nuevaEmpresa.correo ||
-        !this.nuevaEmpresa.sitioWeb
+        !this.nuevaEmpresa.sitioWeb ||
+        !this.nuevaEmpresa.provincia ||
+        !this.nuevaEmpresa.ciudad
       ) {
         alert("Por favor, complete todos los campos antes de guardar.");
         return; // Detener la función si algún campo está vacío
@@ -393,39 +465,62 @@ export default {
         return; // Detener la función si el nombre contiene caracteres no permitidos
       }
 
+      // Validar que la provincia no tenga caracteres raros (solo letras )
+      const provinciaValida = /^[a-zA-Z\s]*$/.test(this.nuevaEmpresa.provincia);
+      if (!provinciaValida) {
+        alert("La provincia solo puede contener letras.");
+        return; // Detener la función si la provincia contiene caracteres no permitidos
+      }
+
+      // Validar que la ciudad no tenga caracteres raros (solo letras )
+      const ciudadValida = /^[a-zA-Z\s]*$/.test(this.nuevaEmpresa.ciudad);
+      if (!ciudadValida) {
+        alert("La ciudad solo puede contener letras.");
+        return; // Detener la función si la ciudad contiene caracteres no permitidos
+      }
+
       // Agregar la nueva empresa a la lista de empresas
       this.empresas.push({
         nom_empresa: this.nuevaEmpresa.nombre,
-        direccion: this.nuevaEmpresa.direccion,
+        nombre_calle: this.nuevaEmpresa.nombre_calle,
         numero_teléfono: this.nuevaEmpresa.telefono,
         correo: this.nuevaEmpresa.correo,
         plan: this.nuevaEmpresa.plan,
         sitio_web: this.nuevaEmpresa.sitioWeb,
+        provincia: this.nuevaEmpresa.provincia,
+        ciudad: this.nuevaEmpresa.ciudad
       });
 
       this.empresasEnviar.push(
         this.nuevaEmpresa.nombre,
-        this.nuevaEmpresa.direccion,
+        this.nuevaEmpresa.nombre_calle,
         this.nuevaEmpresa.telefono,
         this.nuevaEmpresa.correo,
         this.nuevaEmpresa.sitioWeb,
-        this.nuevaEmpresa.plan
+        this.nuevaEmpresa.plan,
+        this.nuevaEmpresa.provincia,
+        this.nuevaEmpresa.ciudad
       );
 
       await insertEmpresa(this.empresasEnviar);
 
-      console.log("Después de agregar a empresas:", this.empresas);
+      // console.log("Después de agregar a empresas:", this.empresas);
 
       // Limpiar el formulario y ocultar el modal
       this.nuevaEmpresa = {
         nombre: "",
-        direccion: "",
+        nombre_calle: "",
         telefono: "",
         correo: "",
         plan: "Standard",
         sitioWeb: "",
+        provincia: "",
+        ciudad: ""
       };
-      this.getEmpresa();
+
+      this.empresasEnviar = [];
+
+      this.getEmpresas();
       this.mostrarModalCrear = false;
     },
 
@@ -433,19 +528,23 @@ export default {
     limpiarFormularioEmpresa() {
       this.nuevaEmpresa = {
         nombre: "",
-        direccion: "",
+        nombre_calle: "",
         telefono: "",
         correo: "",
         plan: "Standard",
         sitioWeb: "",
+        provincia: "",
+        ciudad: ""
       };
 
       // Ocultamos el modal
       this.mostrarModalCrear = false;
     },
 
+    // SELECT EMPRESAS
     async getEmpresas() {
       this.empresas = await getEmpresa();
+      //console.log(this.empresas);
     },
 
     // Método para mostrar los datos de una empresa específica
@@ -460,6 +559,43 @@ export default {
     abrirModalCrearEmpresa() {
       this.mostrarModalCrear = true;
     },
+
+    getChartData() {
+      return {
+        labels: ['Aigües de Barcelona', 'Institut de Ciències del Mar', 'INS mi casa', 'Empresa Carlos Ramirez'],
+        datasets: [
+          {
+            label: 'Plan de la Empresa',
+            backgroundColor: ['#2196F3', '#4CAF50', '#FFC107', '#FF5722'],
+            data: [2, 3, 1, 2] // Se debe adaptar según la cantidad de planes
+          }
+        ]
+      };
+    },
+    getChartOptions() {
+      return {
+        responsive: true,
+        maintainAspectRatio: false,
+        scales: {
+          xAxes: [{
+            ticks: {
+              autoSkip: false
+            }
+          }],
+          yAxes: [{
+            ticks: {
+              beginAtZero: true,
+              stepSize: 1
+            }
+          }]
+        },
+        legend: {
+          display: true,
+          position: 'top'
+        }
+      };
+    }
+   
   },
 
   //CONSOLA
@@ -470,6 +606,9 @@ export default {
   },
   mounted() {
     console.log("MONTADO");
+    this.getEmpresas().then(() => {
+
+    });
   },
 
   updated() {
@@ -479,7 +618,7 @@ export default {
 </script>
 
 <script setup>
-import DefaultBar from "@/components/appbar.vue";
+import DefaultBar from "@/layouts/default/AppBar.vue";
 </script>
 
 <style scoped>
